+++ src/test/java/org/akv/leetcode/practice/easy/two_pointers/TwoPointerProblemTest.java
@@ -1,0 +1,78 @@
+package org.akv.leetcode.practice.easy.two_pointers;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class TwoPointerProblemTest {
+
+    TwoPointerProblem problem = new TwoPointerProblem();
+
+    @Test
+    void shortestDistanceToChar() {
+
+        problem.shortestDistanceToChar("aabaaaaa", 'b');
+
+        problem.shortestDistanceToChar("aaaaaabaaa", 'b');
+        problem.shortestDistanceToChar("abaaaabaaa", 'b');
+        problem.shortestDistanceToChar("abaabaabaaa", 'b');
+        problem.shortestDistanceToChar("loveleetcode", 'e');
+//[3,2,1,0,1,0,0,1,2,2,1,0]
+    }
+
+
+    @Test
+    void findArrConcatValue() {
+
+        assertEquals(247, problem.findArrConcatValue(new int[]{1, 2, 33, 4}));
+
+        assertEquals(673, problem.findArrConcatValue(new int[]{5, 14, 13, 8, 12}));
+
+        assertEquals(585, problem.findArrConcatValue(new int[]{1, 45, 2, 33}));
+        assertEquals(11, problem.findArrConcatValue(new int[]{1, 1}));
+
+
+    }
+
+
+    @Test
+    void firstPalindrome() {
+        assertEquals("", problem.firstPalindrome(new String[]{}));
+
+        assertEquals("s", problem.firstPalindrome(new String[]{"s"}));
+
+        assertEquals("", problem.firstPalindrome(new String[]{"sa", "sb"}));
+        assertEquals("ss", problem.firstPalindrome(new String[]{"sa", "sb", "ss"}));
+
+
+        assertEquals("ada", problem.firstPalindrome(new String[]{"abc", "car", "ada", "racecar", "cool"}));
+    }
+
+    @Test
+    void testCountPairsWithLoop() {
+        assertEquals(3, problem.countPairsWithLoop(new int[]{-1, 1, 2, 3, 1}, 2));
+        assertEquals(0, problem.countPairsWithLoop(new int[]{-1, 1, 2, 3, 1}, -2));
+        assertEquals(10, problem.countPairsWithLoop(new int[]{-6, 2, 5, -2, -7, -1, 3}, -2));
+
+    }
+
+    @Test
+    void arithmeticTriplets() {
+        assertEquals(0, problem.arithmeticTriplets(new int[]{0, 2}, 2));
+        assertEquals(2, problem.arithmeticTriplets(new int[]{0, 1, 4, 6, 7, 10}, 3));
+        assertEquals(2, problem.arithmeticTriplets(new int[]{ 1, 4, 7, 10}, 3));
+        assertEquals(2, problem.arithmeticTriplets(new int[]{4, 5, 6, 7, 8, 9}, 2));
+    }
+
+    @Test
+    void findPerms() {
+        assertEquals(new int[]{0, 4, 1, 3, 2}, problem.findPerms("IDID"));
+    }
+    @Test
+    void sortedSquares() {
+        assertArrayEquals(new int[]{0, 9, 16, 25, 100}, problem.sortedSquares(new int[]{-4, -3, 0, 5, 10}));
+        assertArrayEquals(new int[]{1, 4, 9}, problem.sortedSquares(new int[]{-2, 1, 3}));
+
+    }
+
+}
